Tickets - Fix log message deletion by ensuring log channel exists before fetching message; update version to 2.14.2

+++ tickets/common/utils.py
@@ -787,11 +787,12 @@ async def prune_invalid_tickets(
             log_channel_id = panel["log_channel"]
             if log_channel_id and log_message_id:
                 log_channel = guild.get_channel(log_channel_id)
-                try:
-                    log_message = await log_channel.fetch_message(log_message_id)
-                    await log_message.delete()
-                except (discord.NotFound, discord.Forbidden):
-                    pass
+                if log_channel:
+                    try:
+                        log_message = await log_channel.fetch_message(log_message_id)
+                        await log_message.delete()
+                    except (discord.NotFound, discord.Forbidden):
+                        pass
 
     if users_to_remove or tickets_to_remove:
         async with config.guild(guild).opened() as opened:
